Ignore `timeout` events
An informative timeout event is emitted after a test has been running
for 60 seconds. The test is not stopped, but continues running and will
return its result at a later point in time.

This event was previously unrecognized and would cause cargo2junit to
panic.

Ignore this event for now and leave a comment in code for further action
that might be potentially required in the future.