Handle --enforce-time and --show-output
This change adds support for tests that fail due to timeout (via
Rust's --enforce-time flag). This is achieved by treating a test
event's `stdout` as optional and passing through the contents of
`reason`. Previously such a test event led to a parsing error.

The change also adds support for passing through stdout for passing
tests (generated via Rust's --show-output). Previously this data
was silently dropped.

This raises the question of where to include this new information in
the output XML. For passing tests, the contents of stdout is included
in the `<system-out>` node, via `TestCase::system_out`.

For failing tests, the contents of stdout was previously included
in the `message` XML attribute, via `TestResult::Failed::message`. It
seems more correct to include this as the inner text of the
`<failure>` node and to put `reason` (if it exists) into `message`.
Doing this requires upgrading to junit-result 0.4.1, which (for better
or for worse) changes the target of `TestCase::system_out` for failed
tests from the `<system-out>` node to the `<failure>` node.