Handle test logs starting with {
Don't error out when parsing a line starting with '{' but can't be parsed as JSON test output. Just ignore it the same way we do all lines not starting with '{'. It could instead be the code under test writing logs starting {.

This is normally avoided by splitting cargo test output and test logs, but isn't possible under cargo-llvm-cov which redirects it all to stderr.

Signed-off-by: Nigel Thorpe <nigel.thorpe@metaswitch.com>