Make it possible to set suite name prefix.
In a CI environment where tests for many different packages are executed
and reports are gathered, it becomes difficult to separate them from
each other. For example Jenkins will get confused and try to merge all
reports together.

By allowing the system running the tests to override the prefix it can
use e.g. the package name as a prefix to work around this issue.